Extract composite_type.rs
