+++ pgrx-sql-entity-graph/src/lib.rs
@@ -52,6 +52,7 @@ pub use used_type::{UsedType, UsedTypeEntity};
 
 pub(crate) mod aggregate;
 pub(crate) mod control_file;
+pub(crate) mod composite_type;
 pub(crate) mod enrich;
 pub(crate) mod extension_sql;
 pub(crate) mod extern_args;
